mediatek/config: update Kconfig according to live device
It looks like the configuration options in this repository do not match
what we have been using on our devices (and indeed, building the kernel
with the options currently used in this repository produces a binary
that won't fit in the recovery partition). So, update the options by
alignign them with those found through /proc/config.gz on a live device.

The produced image boots, seems to work fine, and is actually small
enough that it can be put into the recovery partition along with the
recovery binaries.