mediatek: fix usage of ioctls in FM driver
The Linux kernel documentation clearly states that the third parameter
of the ioctl helper macros is the type of the data being passed, not a
pointer to it. Passing a pointer not only causes all ioctls to have the
same data size, but, most importantly, makes it impossible to use the
same userspace code for 32 and 64 bits architectures.

This also aligns the code to that of the M10 tablet.